# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Saving Grains API
  description: |
    The Saving Grains API provides endpoints for user authentication, user management, and wallet request functionality. 

    ## Authentication
    Most endpoints require authentication via JWT token passed in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Error Handling
    All API errors follow a consistent format with an `error` field containing the error message.
  version: 1.0.0
  contact:
    name: Saving Grains Support
    email: support@savinggrains.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.savinggrains.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WalletRequest:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        user:
          type: string
          description: User ID who created the request
          example: "507f1f77bcf86cd799439012"
        amount:
          type: number
          minimum: 1
          example: 100.50
        payment_method:
          type: string
          enum: ["Cash Payment", "Mobile Money", "Bank Transfer"]
          example: "Mobile Money"
        provider:
          type: string
          nullable: true
          example: "MTN"
        phone_number:
          type: string
          nullable: true
          example: "+233123456789"
        bank_name:
          type: string
          nullable: true
          example: "Ghana Commercial Bank"
        branch_name:
          type: string
          nullable: true
          example: "Accra Main Branch"
        reason:
          type: string
          example: "Monthly savings deposit"
        status:
          type: string
          enum: ["pending", "approved", "declined", "successful"]
          default: "pending"
          example: "pending"
        reviewed_by:
          type: string
          nullable: true
          description: User ID who reviewed the request
          example: "507f1f77bcf86cd799439013"
        reviewed_at:
          type: string
          format: date-time
          nullable: true
          example: "2023-01-01T00:00:00.000Z"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message description
          example: "Invalid credentials"
...
    CreateWalletRequestRequest:
      type: object
      required:
        - amount
        - payment_method
        - reason
      properties:
        amount:
          type: number
          minimum: 1
          description: Amount to be added to wallet
          example: 100.50
        payment_method:
          type: string
          enum: ["Cash Payment", "Mobile Money", "Bank Transfer"]
          description: Method of payment for the wallet request
          example: "Mobile Money"
        provider:
          type: string
          description: Mobile money provider (required for Mobile Money payments)
          example: "MTN"
        phone_number:
          type: string
          description: Phone number for mobile money (required for Mobile Money payments)
          example: "+233123456789"
        bank_name:
          type: string
          description: Name of the bank (required for Bank Transfer payments)
          example: "Ghana Commercial Bank"
        branch_name:
          type: string
          description: Bank branch name (required for Bank Transfer payments)
          example: "Accra Main Branch"
        reason:
          type: string
          description: Reason for the wallet funding request
          example: "Monthly savings deposit"
paths:
...
  /wallet-request:
    post:
      summary: Create a wallet funding request
      description: |
        Creates a new request to add funds to the user's wallet through various payment methods.

        **Field Requirements by Payment Method:**
        - **Cash Payment:** amount, payment_method, reason
        - **Mobile Money:** amount, payment_method, provider, phone_number, reason
        - **Bank Transfer:** amount, payment_method, bank_name, branch_name, reason
      tags:
        - Wallet Requests
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletRequestRequest"
            examples:
              cash_payment:
                summary: Cash Payment Example
                value:
                  amount: 100.00
                  payment_method: "Cash Payment"
                  reason: "Monthly savings deposit"
              mobile_money:
                summary: Mobile Money Example
                value:
                  amount: 250.50
                  payment_method: "Mobile Money"
                  provider: "MTN"
                  phone_number: "+233123456789"
                  reason: "Emergency fund deposit"
              bank_transfer:
                summary: Bank Transfer Example
                value:
                  amount: 500.00
                  payment_method: "Bank Transfer"
                  bank_name: "Ghana Commercial Bank"
                  branch_name: "Accra Main Branch"
                  reason: "Bulk savings transfer"
      responses:
        "201":
          description: Wallet request created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletRequest"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Missing required fields"
        "401":
          description: Unauthorized - invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Failed to create wallet request"
security:
  - bearerAuth: []